volume/tone manager
