+++ apps/settings/js/panels/sound/panel.js
@@ -5,14 +5,19 @@ define(function(require) {
   'use strict';
 
   var SettingsPanel = require('modules/settings_panel');
-  var Sound = require('panels/sound/sound');
+  var VolumeManager = require('panels/sound/volume_manager');
+  var ToneManager = require('panels/sound/tone_manager');
 
   return function ctor_sound_panel() {
-    var sound = new Sound();
+    var volumeManager = VolumeManager();
+    var toneManager = ToneManager();
 
     return SettingsPanel({
       onInit: function(panel) {
-        var elements = {
+        var sliders_element = panel.querySelectorAll('.volume input');
+        volumeManager.init(sliders_element);
+
+        var tm_elements = {
           toneSelector: panel.querySelector('.touch-tone-selector'),
           alertToneSelection: panel.querySelector('.alert-tone-selection'),
           ringToneSelection: panel.querySelector('.ring-tone-selection'),
@@ -20,7 +25,7 @@ define(function(require) {
           vibrationSetting: panel.querySelector('.vibration-setting'),
           manageTones: panel.querySelector('.manage-tones-button')
         };
-        sound.init(elements);
+        toneManager.init(tm_elements);
       }
     });
   };
